refactor(utils): add hasher pkg and add more uts for toml
